Move GetRpcConfig to Job
