+++ CheckerApi/Jobs/Job.cs
@@ -1,7 +1,10 @@
-﻿using System;
+﻿using CheckerApi.Models.Config;
+using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
+using System.Net;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Quartz;
@@ -39,6 +42,19 @@ namespace CheckerApi.Jobs
                     logger.LogInformation($"{GetType().Name} Finished in {elapsed.TotalSeconds} sec.");
                 }
             });
+        }
+        public static RpcConfig GetRpcConfig (IConfiguration config)
+        {
+            return new RpcConfig()
+            {
+                Url = config.GetValue<string>("Node:Url"),
+                Port = config.GetValue<int>("Node:RpcPort"),
+                Credentials = new NetworkCredential()
+                {
+                    UserName = config.GetValue<string>("Node:RpcUser"),
+                    Password = config.GetValue<string>("Node:RpcPass")
+                }
+            };
         }
     }
 }
