+++ CheckerApi/Jobs/NodeJob.cs
@@ -28,7 +28,7 @@ namespace CheckerApi.Jobs
             var logger = serviceProvider.GetService<ILogger<NodeJob>>();
             var mapper = serviceProvider.GetService<IMapper>();
 
-            var rpcConfig = JobCommon.GetRpcConfig(config);
+            var rpcConfig = GetRpcConfig(config);
 
             var cache = serviceProvider.GetService<IMemoryCache>();
             var difficultyResult = dataExtractor.RpcCall(rpcConfig, "getdifficulty");
